REF: move implementations to mixin
